Renames, modules private
